// generated/src/app/api-models/apc-interface-rules.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ApcInterfaceRules {
  lineOfBusiness: string;
  dispositionLevel: string;
  claimAction: string;
  reasonCode: string;
  securityCode: string;
  insertUser: string;
  insertProcess: string;
  insertDatetime: Date;
  updateUser: string;
  updateProcess: string;
  updateDatetime: Date;
}